requirements_production.txt
git+https://github.com/botswana-harvard/django-crypto-fields.git@0.1.14#egg=django-crypto-fields
git+https://github.com/botswana-harvard/django-revision.git@0.1.14#egg=django-revision
git+https://github.com/botswana-harvard/edc-appointment.git@0.2.20#egg=edc-appointment
git+https://github.com/botswana-harvard/edc-base.git@0.2.26#egg=edc-base
git+https://github.com/botswana-harvard/edc-constants.git@0.2.9#egg=edc-constants
git+https://github.com/botswana-harvard/edc-dashboard.git@0.3.2#egg=edc-dashboard
git+https://github.com/botswana-harvard/edc-data-manager.git@0.1.7#egg=edc-data-manager
git+https://github.com/botswana-harvard/edc-device.git@0.2.7#egg=edc-device
git+https://github.com/botswana-harvard/edc-facility.git@0.1.4#egg=edc-facility
git+https://github.com/botswana-harvard/edc-fieldsets.git@0.1.2#egg=edc-fieldsets
git+https://github.com/botswana-harvard/edc-form-validators.git@0.1.3#egg=edc-form-validators
git+https://github.com/botswana-harvard/edc-identifier.git@0.2.12#egg=edc-identifier
git+https://github.com/botswana-harvard/edc-lab.git@0.2.21#egg=edc-lab
git+https://github.com/botswana-harvard/edc-label.git@0.1.9#egg=edc-label
git+https://github.com/botswana-harvard/edc-model-admin.git@0.1.2#egg=edc-model-admin
git+https://github.com/botswana-harvard/edc-model-wrapper.git@0.1.7#egg=edc-model-wrapper
git+https://github.com/botswana-harvard/edc-protocol.git@0.1.5#egg=edc-protocol
git+https://github.com/botswana-harvard/edc-search.git@0.1.6#egg=edc-search
git+https://github.com/botswana-harvard/edc-sms.git@0.1.1#egg=edc-sms
git+https://github.com/botswana-harvard/edc-subject-dashboard.git@0.2.3#egg=edc-subject-dashboard
git+https://github.com/botswana-harvard/edc-metadata.git@0.2.13#egg=edc-metadata
git+https://github.com/botswana-harvard/edc-navbar.git@0.1.5#egg=edc-navbar
git+https://github.com/botswana-harvard/edc-reference.git@0.1.14#egg=edc-reference
git+https://github.com/botswana-harvard/edc-visit-schedule.git@0.2.12#egg=edc-visit-schedule
git+https://github.com/botswana-harvard/cms-dashboard.git@0.1.1#egg=cms-dashboard
git+https://github.com/Botswana-Harvard-Utility-Systems/procurement.git@0.1.1#egg=procurement
git+https://github.com/Botswana-Harvard-Utility-Systems/procurement-dashboard.git@0.1.1#egg=procurement-dashboard
git+https://github.com/Botswana-Harvard-Utility-Systems/bhp-personnel.git@0.1.3#egg=bhp-personnel
git+https://github.com/Botswana-Harvard-Utility-Systems/bhp-personnel-dashboard.git@0.1.1#egg=bhp-personnel-dashboard
git+https://github.com/Botswana-Harvard-Utility-Systems/document-tracking-dashboard.git@0.1.1#egg=document-tracking-dashboard
git+https://github.com/Botswana-Harvard-Utility-Systems/document-tracking.git@0.1.1#egg=document-tracking
git+https://github.com/Botswana-Harvard-Utility-Systems/timesheet.git@0.1.3#egg=timesheet
git+https://github.com/Botswana-Harvard-Utility-Systems/timesheet-dashboard.git@0.1.3#egg=timesheet-dashboard
--pre xhtml2pdf
django-admin-list-filter-dropdown
pdf2image
django_q
xlwt
xhtml2pdf
mysqlclient
pandas
django-crispy-forms
